Saving before wiping

+++ Ticker.Entities/IStock.cs
@@ -1,10 +1,8 @@
-using System;
-
 namespace Ticker.Entities
 {
     public interface IStock
     {
         string Name { get; }
-        double Value { get; set; }
+        decimal Value { get; set; }
     }
 }
